글 작성자만 수정,삭제 가능
글 작성자만 게시물 수정/삭제 가능하도록 구현 완료

+++ src/main/reactjs/src/components/ReadRecruitComponent.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import RecruitService from '../service/RecruitService'
 import './css/Recruit.css';
 
-import ChatPage from '../pages/ChatPage'
+import UserService from '../service/UserService'
 
 class ReadRecruitComponent extends Component {
     constructor(props) {
@@ -12,7 +12,8 @@ class ReadRecruitComponent extends Component {
         this.state = {
             recruit_id : this.props.match.params.recruit_id,
             recruit: {},
-            chatCondition : false
+            chatCondition : false,
+            roleCondition: false,
         }
     }
 
@@ -20,7 +21,33 @@ class ReadRecruitComponent extends Component {
         componentDidMount() {
             RecruitService.getOneRecruit(this.state.recruit_id).then((res) => {
                 this.setState({ recruit: res.data });
+                this.getRoleCondition(this.state.recruit.user_id);
             });
+
+            // UserService에서 user_id 가져오기
+            UserService.getUserName().then(res => {
+                this.setState({
+                    user_id: res.data.id,
+                })
+
+                this.getRoleCondition(this.state.user_id); // 로그인 한 user_id 검사 위해
+            })
+        }
+
+        // 수정, 삭제 권한 있는 지 검사
+        getRoleCondition(write_id, login_id) {
+            if(this.state.recruit.user_id == this.state.user_id) {
+                this.setState({
+                    roleCondition: true,
+                })
+            }
+            else {
+                this.setState({
+                    roleCondition: false,
+                })
+            }
+
+            console.log(this.state.roleCondition);
         }
 
         // 파라미터 값에 따라 페이지에 표시할 내용 변경
@@ -44,7 +71,6 @@ class ReadRecruitComponent extends Component {
             )
         }
 
-
         /*
         returnDate(cTime, uTime) {
             return (
@@ -62,28 +88,40 @@ class ReadRecruitComponent extends Component {
 
         // 글 수정으로 이동
         goToUpdate = (event) => {
+            if(this.state.roleCondition) {
                 event.preventDefault();
                 this.props.history.push(`/createRecruit/${this.state.recruit_id}`);
+            }
+            else {
+                alert("※ 본인만 글을 수정할 수 있습니다");
+            }
         }
 
         deleteView = async function () {
-            if(window.confirm("정말로 글을 삭제하시겠습니까?\n삭제된 글은 복구 할 수 없습니다.")) {
-                RecruitService.deleteRecruit(this.state.recruit_id).then( res => {
-                    console.log("delete result => " + JSON.stringify(res));
+            if(this.state.roleCondition) {
+                if(window.confirm("정말로 글을 삭제하시겠습니까?\n삭제된 글은 복구 할 수 없습니다.")) {
+                    RecruitService.deleteRecruit(this.state.recruit_id).then( res => {
+                        console.log("delete result => " + JSON.stringify(res));
 
-                    if(res.status == 200) {
-                        this.props.history.push('/recruit');
-                    }
-                    else {
-                        alert("글 삭제가 실패했습니다.");
-                    }
-                });
+                        if(res.status == 200) {
+                            this.props.history.push('/recruit');
+                        }
+                        else {
+                            alert("글 삭제를 실패했습니다.");
+                        }
+                    });
+                }
+            }
+
+            else {
+                alert("※ 본인만 글을 삭제할 수 있습니다");
             }
         }
 
     render() {
         return (
             <div>
+            글 작성자는 {this.state.recruit.user_id}
                 <div className = "card col-md-6 offset-md-3">
                     <h3 className ="text-center"> Read Detail
                         <button className="btn btn-info"
@@ -111,9 +149,10 @@ class ReadRecruitComponent extends Component {
                     </div>
                 </div>
                 <div>
-                <div className="chatComponent">
+                {/* <div className="chatComponent">
                     { this.state.chatCondition ? <ChatPage/> : null }
-                </div></div>
+                </div> */}
+                </div>
             </div>
         );
     }
